agregar validaciones

+++ rockets/keppler_capsules/app/models/keppler_capsules/capsule.rb
@@ -3,15 +3,19 @@ module KepplerCapsules
   class Capsule < ActiveRecord::Base
     include ActivityHistory
     include CloneRecord
+    include KepplerCapsules::Concerns::ActionsOnDatabase
+    include KepplerCapsules::Concerns::StringActions
     require 'csv'
     acts_as_list
     has_many :capsule_fields, dependent: :destroy, inverse_of: :capsule
+    has_many :capsule_validations, dependent: :destroy, inverse_of: :capsule
     before_destroy :uninstall
     accepts_nested_attributes_for :capsule_fields, reject_if: :all_blank, allow_destroy: true
+    accepts_nested_attributes_for :capsule_validations, reject_if: :all_blank, allow_destroy: true
 
     validates_presence_of :name
     validates_uniqueness_of :name
-    before_validation :convert_to_downcase, :without_special_characters
+    before_validation :without_special_characters
 
 
     # Fields for the search form in the navbar
@@ -43,14 +47,17 @@ module KepplerCapsules
     end
 
     def install
+      return if table_exists?("keppler_capsules_#{self.name}")
       fields = self.capsule_fields.map { |f| "#{f.name_field}:#{f.format_field}" }
       fields = fields.join(' ')
       system("cd rockets/keppler_capsules && rails g keppler_capsule_scaffold #{self.name} #{fields} position:integer deleted_at:datetime:index")
+      create_pg_table(self.name)
       system('rake keppler_capsules:install:migrations')
       system('rake db:migrate')
     end
 
     def uninstall
+      return unless table_exists?("keppler_capsules_#{self.name}")
       system("cd rockets/keppler_capsules && rails d keppler_capsule_scaffold #{self.name}")
       FileUtils.rm_rf("#{url_capsule}/app/views/keppler_capsules/admin/#{self.name}")
       delete_pg_table("keppler_capsules_#{self.name}")
@@ -61,61 +68,36 @@ module KepplerCapsules
     def new_attributes(table, attributes)
       attributes.each do |key, value|
         if value[:name_field]
-          add_field_pg_table(value, table)
+          field = CapsuleField.where(name_field: value[:name_field])
+          if field.count == 1
+            add_field_pg_table(value, table)
+          else
+            field.last.delete
+          end
         end
       end
       system('rake keppler_capsules:install:migrations')
       system('rake db:migrate')
     end
 
+    def new_validations(table, attributes)
+      attributes.each do |key, value|
+        if value[:name]
+          validation = CapsuleValidation.where(name: value[:name], field: value[:field])
+          add_validation_to(table, value) if validation.count == 1
+        end
+      end
+    end
+
     private
-
-    def url_capsule
-      "#{Rails.root}/rockets/keppler_capsules"
-    end
-
-    def convert_to_downcase
-      self.name.pluralize.downcase!
-    end
 
     def without_special_characters
       self.name.gsub!(' ', '_')
       special_characters.each { |sc| self.name.gsub!(sc, '') }
     end
 
-    def special_characters
-      [
-        '/', '.', '@', '"', "'", '%', '&', '$',
-        '?', '¿', '/', '=', ')', '(', '#', '{',
-        '}', ',', ';', ':', '[', ']', '^', '`',
-        '¨', '~', '+', '-', '*', '¡', '!', '|',
-        '¬', '°', '<', '>', '·', '½'
-      ]
-    end
-
-    def delete_pg_table(table)
-      system("cd rockets/keppler_capsules && rails g migration Drop#{table.split('_').map(&:capitalize).join('')}")
-      migration = Dir.entries("#{url_capsule}/db/migrate").sort.last
-      out_file = File.open("#{url_capsule}/db/migrate/#{migration}", "w")
-      out_file.puts(migrate_delete_format(table));
-      out_file.close
-    end
-
-    def add_field_pg_table(column, table)
-      system("cd rockets/keppler_capsules && rails g migration add_#{column[:name_field]}_to_keppler_capsules_#{table} #{column[:name_field]}:#{column[:format_field]}")
-    end
-
-    def delete_field_pg_table(column, table)
-    end
-
-    def migrate_delete_format(table)
-      [
-        "class Drop#{table.split('_').map(&:capitalize).join('')} < ActiveRecord::Migration[5.2]\n",
-        "  def change\n",
-        "    drop_table :#{table}\n",
-        "  end\n",
-        "end"
-      ].join
+    def table_exists?(table)
+      ActiveRecord::Base.connection.table_exists?(table)
     end
 
   end
